feat : UserController 토큰검사 - 테스트 코드도 같이 변경

+++ server/src/main/java/project/server/domain/user/UserService.java
@@ -47,7 +47,8 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(long userId) {
+    public void deleteUser(long userId, Authentication authentication) {
+        if(authentication == null) throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_USER);
         User user = findUser(userId);
         userRepository.delete(user);
     }
